+++ src/components/ContainerLogin/page.module.css
@@ -96,240 +96,7 @@
   letter-spacing: -0.6px;
 }
 
-.containerRecuperarSenha {
-  display: flex;
-  width: 100%;
-  height: 100vh;
-  justify-content: space-between;
 
-}
-
-.containerForm {
-  background: #121212;
-  color: white;
-  width: 40%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  
-}
-
-.subContainer {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-  gap: 10px;
-  margin: 0 auto;
-  padding: 10px;
-}
-
-.logoContainer {
-  width: 60px;
-  height: 40px;
-}
-
-.titleContainerForm {
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 32px; /* 177.778% */
-  letter-spacing: -1.08px;
-  }
-
-.titleForm {
-  margin-top: 100px;
-  font-size: 30px;
-  font-weight: 500;
-  line-height: 29.05px;
-  letter-spacing: -0.02em;
-  text-align: left;
-}
-
-.paragraphForm {
-  color: #787878;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 20px;
-  text-align: left;
-  margin-bottom: 50px;
-  margin-top: 20px;
-}
-
-.formField {
-  display: flex;
-  margin-left: 40px;
-}
-
-.inputNumber {
-  margin: 10px;
-  background-color: #121212;
-  border: 1px solid #3b3b3b;
-  padding: 10px;
-  border-radius: 12px;
-  margin-bottom: 48px;
-  width: 50px;
-  height: 50px;
-
-  color: white;
-  font-size: 24px;
-  font-weight: 600;
-}
-
-.formField input[type="text"] {
-  border: 1px solid #3b3b3b;
-  padding: 10px; 
-  outline: none;
-}
-
-.formField input[type="text"]:focus {
-  border: 1px solid red;
-}
-
-
-.buttonRecuperar {
-  width: 366px;
-  height: 51px;
-  padding: 16px 10px 16px 10px;
-  gap: 10px;
-  border-radius: 12px;
-  margin-bottom: 20px;
-  margin-top: 20px;
-  margin-left: 80px;
-  background: #eb4242;
-  color: white;
-  border: none;
-  
-  font-size: 16px;
-  font-weight: 600;
-  line-height: 19.36px;
-
-  display: flex;
-  justify-content: center;
-}
-
-.buttonRecuperar:hover {
-  background: #872525;
-  cursor: pointer;
-}
-
-.buttonLogin {
-  width: 366px;
-  height: 51px;
-  padding: 16px 10px 16px 10px;
-  gap: 10px;
-  border-radius: 12px;
-  margin-bottom: 20px;
-  margin-top: 20px;
-  margin-left: 80px;
-  background: #3b3b3b;
-  color: white;
-  border: none;
-
-  font-size: 16px;
-  font-weight: 600;
-  line-height: 19.36px;
-
-  display: flex;
-  justify-content: center;
-}
-
-.buttonLogin:hover {
-  background: #121212;
-  cursor: pointer;
-}
-
-
-.submitContainer{
-  display: flex;
-  flex-direction: column;
-  gap:10px  
-}
-
- /* Estilos gerais para telas menores */
-@media (max-width: 768px) {
-
-  .containerRecuperarSenha {
-    flex-direction: column; /* Coloca os elementos um abaixo do outro */
-    justify-content: space-between; /* Centraliza os elementos na tela */
-    align-items: flex-start;
-    height: 100vh;
-  }
-
-  .containerForm {
-    width: 100%;
-    height: 100%;
-    margin-top: 30px;
-    
-  }
-  .form{
-    width: 85%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-   
-  }
-  .subContainer{
-    margin-top: 50px;
-    margin-bottom: -30px;
-    align-items: center;
-    justify-content: center;
-  }
-  .titleForm{
-    font-size: 20px;
-    margin-bottom: -20px;
-   }
-
-   .titleContainerForm{
-    font-size: 16px;
-
-  }
-
-  .paragraphForm{
-  
-    font-size: 14px;
-    margin-bottom: 10px;
-  }
-  .paragraphField{
-    font-size: 12px;
-    
-  }
-  .formField {
-    width: 100%;
-    align-self: center;
-    justify-content: center;
-    align-items: center;
-    margin: 0 auto;
-    margin-bottom: 5px;
-    
-  }
-
-  .inputNumber {
-    width: 36px;
-    height: 36px;
-    font-size: 18px;
-    margin: 3px;
-  }
-
-  .buttonRecuperar, .buttonLogin {
-    width: 70%;
-    margin: 10px 0;
-  }
-
-  .buttonLogin {
-    margin-top: 10px;
-  }
-
-  .submitContainer {
-    margin-top: 20px;
-    align-items: center;
-  }
-  .alignContainer{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-}
 
 /* Tela pequena (mobile) */
 @media (max-width: 768px) {
